+++ TODO.txt
@@ -1,0 +1,3 @@
+1. deviceType
+两类：温度(10)和电表(20)
+
