+++ monitor.cfg
@@ -1,4 +1,5 @@
-IP=127.0.0.1
+#update key: PORT
+#Fri Nov 10 16:58:23 CST 2017
 PORT=2404
 ALARM_TIMER=
-
+IP=127.0.0.1
