--- src/main/java/core/CallBack.java
@@ -1,5 +1,0 @@
-package core;
-
-public interface CallBack {
-    public void method(Object arg);
-}
