+++ src/main/java/core/defs/DeviceType.java
@@ -1,0 +1,11 @@
+package core.defs;
+
+public enum DeviceType {
+    DEVICE_HUMITURE(10),
+    DEVICE_EMETER(20);
+
+    int type;
+    DeviceType(int type) {
+        this.type = type;
+    }
+}
