+++ src/main/java/core/defs/ErrorMsg.java
@@ -1,0 +1,5 @@
+package core.defs;
+
+public class ErrorMsg {
+    public static String UNKNOWN_DEVICE_TYPE = "Unknown device type (%d).";
+}
