+++ src/main/java/core/event/MCallback.java
@@ -1,0 +1,7 @@
+package core.event;
+
+import core.event.MEvent;
+
+public interface MCallback {
+    public void handleEvent(MEvent event);
+}
