+++ src/main/java/core/event/MEvent.java
@@ -1,15 +1,11 @@
 package core.event;
 
-import core.CallBack;
-
-import java.util.Vector;
-
 public class MEvent {
     private MType type;
     private Object[] argArray;
-    private CallBack handler;
+    private MCallback handler;
 
-    public MEvent(MType type, Object[] argArray, CallBack handler) {
+    public MEvent(MType type, Object[] argArray, MCallback handler) {
         this.type = type;
         this.argArray = argArray;
         this.handler = handler;
@@ -19,4 +15,11 @@ public class MEvent {
         return type;
     }
 
+    public Object[] getArgArray() {
+        return argArray;
+    }
+
+    public MCallback getHandler() {
+        return handler;
+    }
 }
