+++ src/main/java/core/event/MType.java
@@ -1,7 +1,7 @@
 package core.event;
 
 public enum MType {
-    ID_DEIVCE_DATA,
+    ID_DEVICE_DATA,
     ID_ALARM_TIMER,
     ID_UPDATE_DATA,
     ID_CHECK_DEVICE_STATUS,
