+++ src/main/java/gui/dialog/SettingDialog.java
@@ -1,35 +1,32 @@
 package gui.dialog;
 
 import core.Main;
-import gui.pane.DevicePane;
+import util.PropertiesUtil;
 
 import javax.swing.*;
-import java.awt.*;
 import java.awt.event.*;
-import java.util.Vector;
 
-public class DeviceDialog extends JDialog {
+public class SettingDialog extends JDialog {
     private JPanel contentPane;
     private JButton buttonOK;
     private JButton buttonCancel;
-    private JPanel pane;
-    private DevicePane devPane;
+    private JTabbedPane tabbedPane1;
+    private JTextField txtIp;
+    private JTextField txtPort;
 
-    public DeviceDialog() {
+    public SettingDialog() {
         super(Main.getMainWindow(), true);
 
         setContentPane(contentPane);
         getRootPane().setDefaultButton(buttonOK);
 
         buttonOK.addActionListener(new ActionListener() {
-            @Override
             public void actionPerformed(ActionEvent e) {
                 onOK();
             }
         });
 
         buttonCancel.addActionListener(new ActionListener() {
-            @Override
             public void actionPerformed(ActionEvent e) {
                 onCancel();
             }
@@ -38,7 +35,6 @@ public class DeviceDialog extends JDialog {
         // call onCancel() when cross is clicked
         setDefaultCloseOperation(DO_NOTHING_ON_CLOSE);
         addWindowListener(new WindowAdapter() {
-            @Override
             public void windowClosing(WindowEvent e) {
                 onCancel();
             }
@@ -46,21 +42,26 @@ public class DeviceDialog extends JDialog {
 
         // call onCancel() on ESCAPE
         contentPane.registerKeyboardAction(new ActionListener() {
-            @Override
             public void actionPerformed(ActionEvent e) {
                 onCancel();
             }
         }, KeyStroke.getKeyStroke(KeyEvent.VK_ESCAPE, 0), JComponent.WHEN_ANCESTOR_OF_FOCUSED_COMPONENT);
 
-        // set device pane
-        devPane = new DevicePane();
-        pane.setLayout(new BorderLayout());
-        pane.add(devPane.contentPane);
 
+        // load config data
+        String ip = PropertiesUtil.getInstance().getValue("IP");
+        String port = PropertiesUtil.getInstance().getValue("PORT");
+        txtIp.setText(ip);
+        txtPort.setText(port);
     }
 
     private void onOK() {
         // add your code here
+        String ip = txtIp.getText().trim();
+        String port = txtPort.getText().trim();
+        PropertiesUtil.getInstance().store("IP", ip);
+        PropertiesUtil.getInstance().store("PORT", port);
+
         dispose();
     }
 
@@ -70,17 +71,9 @@ public class DeviceDialog extends JDialog {
     }
 
     public static void main(String[] args) {
-        DeviceDialog dialog = new DeviceDialog();
+        SettingDialog dialog = new SettingDialog();
         dialog.pack();
         dialog.setVisible(true);
         System.exit(0);
-    }
-
-    public void fillDevPaneData(Vector<Object> data) {
-        devPane.populate(data);
-    }
-
-    public void freezeDevPaneData() {
-        devPane.freeze();
     }
 }
