--- src/mapper/DeviceMapper.java
@@ -1,30 +1,0 @@
-package mapper;
-
-import java.util.List;
-import org.apache.ibatis.annotations.Param;
-import po.Device;
-import po.DeviceExample;
-
-public interface DeviceMapper {
-    long countByExample(DeviceExample example);
-
-    int deleteByExample(DeviceExample example);
-
-    int deleteByPrimaryKey(Integer code);
-
-    int insert(Device record);
-
-    int insertSelective(Device record);
-
-    List<Device> selectByExample(DeviceExample example);
-
-    Device selectByPrimaryKey(Integer code);
-
-    int updateByExampleSelective(@Param("record") Device record, @Param("example") DeviceExample example);
-
-    int updateByExample(@Param("record") Device record, @Param("example") DeviceExample example);
-
-    int updateByPrimaryKeySelective(Device record);
-
-    int updateByPrimaryKey(Device record);
-}
