--- src/po/DeviceDataExample.java
@@ -1,441 +1,0 @@
-package po;
-
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-
-public class DeviceDataExample {
-    protected String orderByClause;
-
-    protected boolean distinct;
-
-    protected List<Criteria> oredCriteria;
-
-    public DeviceDataExample() {
-        oredCriteria = new ArrayList<Criteria>();
-    }
-
-    public void setOrderByClause(String orderByClause) {
-        this.orderByClause = orderByClause;
-    }
-
-    public String getOrderByClause() {
-        return orderByClause;
-    }
-
-    public void setDistinct(boolean distinct) {
-        this.distinct = distinct;
-    }
-
-    public boolean isDistinct() {
-        return distinct;
-    }
-
-    public List<Criteria> getOredCriteria() {
-        return oredCriteria;
-    }
-
-    public void or(Criteria criteria) {
-        oredCriteria.add(criteria);
-    }
-
-    public Criteria or() {
-        Criteria criteria = createCriteriaInternal();
-        oredCriteria.add(criteria);
-        return criteria;
-    }
-
-    public Criteria createCriteria() {
-        Criteria criteria = createCriteriaInternal();
-        if (oredCriteria.size() == 0) {
-            oredCriteria.add(criteria);
-        }
-        return criteria;
-    }
-
-    protected Criteria createCriteriaInternal() {
-        Criteria criteria = new Criteria();
-        return criteria;
-    }
-
-    public void clear() {
-        oredCriteria.clear();
-        orderByClause = null;
-        distinct = false;
-    }
-
-    protected abstract static class GeneratedCriteria {
-        protected List<Criterion> criteria;
-
-        protected GeneratedCriteria() {
-            super();
-            criteria = new ArrayList<Criterion>();
-        }
-
-        public boolean isValid() {
-            return criteria.size() > 0;
-        }
-
-        public List<Criterion> getAllCriteria() {
-            return criteria;
-        }
-
-        public List<Criterion> getCriteria() {
-            return criteria;
-        }
-
-        protected void addCriterion(String condition) {
-            if (condition == null) {
-                throw new RuntimeException("Value for condition cannot be null");
-            }
-            criteria.add(new Criterion(condition));
-        }
-
-        protected void addCriterion(String condition, Object value, String property) {
-            if (value == null) {
-                throw new RuntimeException("Value for " + property + " cannot be null");
-            }
-            criteria.add(new Criterion(condition, value));
-        }
-
-        protected void addCriterion(String condition, Object value1, Object value2, String property) {
-            if (value1 == null || value2 == null) {
-                throw new RuntimeException("Between values for " + property + " cannot be null");
-            }
-            criteria.add(new Criterion(condition, value1, value2));
-        }
-
-        public Criteria andCodeIsNull() {
-            addCriterion("code is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andCodeIsNotNull() {
-            addCriterion("code is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andCodeEqualTo(Integer value) {
-            addCriterion("code =", value, "code");
-            return (Criteria) this;
-        }
-
-        public Criteria andCodeNotEqualTo(Integer value) {
-            addCriterion("code <>", value, "code");
-            return (Criteria) this;
-        }
-
-        public Criteria andCodeGreaterThan(Integer value) {
-            addCriterion("code >", value, "code");
-            return (Criteria) this;
-        }
-
-        public Criteria andCodeGreaterThanOrEqualTo(Integer value) {
-            addCriterion("code >=", value, "code");
-            return (Criteria) this;
-        }
-
-        public Criteria andCodeLessThan(Integer value) {
-            addCriterion("code <", value, "code");
-            return (Criteria) this;
-        }
-
-        public Criteria andCodeLessThanOrEqualTo(Integer value) {
-            addCriterion("code <=", value, "code");
-            return (Criteria) this;
-        }
-
-        public Criteria andCodeIn(List<Integer> values) {
-            addCriterion("code in", values, "code");
-            return (Criteria) this;
-        }
-
-        public Criteria andCodeNotIn(List<Integer> values) {
-            addCriterion("code not in", values, "code");
-            return (Criteria) this;
-        }
-
-        public Criteria andCodeBetween(Integer value1, Integer value2) {
-            addCriterion("code between", value1, value2, "code");
-            return (Criteria) this;
-        }
-
-        public Criteria andCodeNotBetween(Integer value1, Integer value2) {
-            addCriterion("code not between", value1, value2, "code");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam1IsNull() {
-            addCriterion("param1 is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam1IsNotNull() {
-            addCriterion("param1 is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam1EqualTo(Float value) {
-            addCriterion("param1 =", value, "param1");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam1NotEqualTo(Float value) {
-            addCriterion("param1 <>", value, "param1");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam1GreaterThan(Float value) {
-            addCriterion("param1 >", value, "param1");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam1GreaterThanOrEqualTo(Float value) {
-            addCriterion("param1 >=", value, "param1");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam1LessThan(Float value) {
-            addCriterion("param1 <", value, "param1");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam1LessThanOrEqualTo(Float value) {
-            addCriterion("param1 <=", value, "param1");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam1In(List<Float> values) {
-            addCriterion("param1 in", values, "param1");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam1NotIn(List<Float> values) {
-            addCriterion("param1 not in", values, "param1");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam1Between(Float value1, Float value2) {
-            addCriterion("param1 between", value1, value2, "param1");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam1NotBetween(Float value1, Float value2) {
-            addCriterion("param1 not between", value1, value2, "param1");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam2IsNull() {
-            addCriterion("param2 is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam2IsNotNull() {
-            addCriterion("param2 is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam2EqualTo(Float value) {
-            addCriterion("param2 =", value, "param2");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam2NotEqualTo(Float value) {
-            addCriterion("param2 <>", value, "param2");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam2GreaterThan(Float value) {
-            addCriterion("param2 >", value, "param2");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam2GreaterThanOrEqualTo(Float value) {
-            addCriterion("param2 >=", value, "param2");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam2LessThan(Float value) {
-            addCriterion("param2 <", value, "param2");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam2LessThanOrEqualTo(Float value) {
-            addCriterion("param2 <=", value, "param2");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam2In(List<Float> values) {
-            addCriterion("param2 in", values, "param2");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam2NotIn(List<Float> values) {
-            addCriterion("param2 not in", values, "param2");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam2Between(Float value1, Float value2) {
-            addCriterion("param2 between", value1, value2, "param2");
-            return (Criteria) this;
-        }
-
-        public Criteria andParam2NotBetween(Float value1, Float value2) {
-            addCriterion("param2 not between", value1, value2, "param2");
-            return (Criteria) this;
-        }
-
-        public Criteria andRecordtimeIsNull() {
-            addCriterion("recordTime is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andRecordtimeIsNotNull() {
-            addCriterion("recordTime is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andRecordtimeEqualTo(Date value) {
-            addCriterion("recordTime =", value, "recordtime");
-            return (Criteria) this;
-        }
-
-        public Criteria andRecordtimeNotEqualTo(Date value) {
-            addCriterion("recordTime <>", value, "recordtime");
-            return (Criteria) this;
-        }
-
-        public Criteria andRecordtimeGreaterThan(Date value) {
-            addCriterion("recordTime >", value, "recordtime");
-            return (Criteria) this;
-        }
-
-        public Criteria andRecordtimeGreaterThanOrEqualTo(Date value) {
-            addCriterion("recordTime >=", value, "recordtime");
-            return (Criteria) this;
-        }
-
-        public Criteria andRecordtimeLessThan(Date value) {
-            addCriterion("recordTime <", value, "recordtime");
-            return (Criteria) this;
-        }
-
-        public Criteria andRecordtimeLessThanOrEqualTo(Date value) {
-            addCriterion("recordTime <=", value, "recordtime");
-            return (Criteria) this;
-        }
-
-        public Criteria andRecordtimeIn(List<Date> values) {
-            addCriterion("recordTime in", values, "recordtime");
-            return (Criteria) this;
-        }
-
-        public Criteria andRecordtimeNotIn(List<Date> values) {
-            addCriterion("recordTime not in", values, "recordtime");
-            return (Criteria) this;
-        }
-
-        public Criteria andRecordtimeBetween(Date value1, Date value2) {
-            addCriterion("recordTime between", value1, value2, "recordtime");
-            return (Criteria) this;
-        }
-
-        public Criteria andRecordtimeNotBetween(Date value1, Date value2) {
-            addCriterion("recordTime not between", value1, value2, "recordtime");
-            return (Criteria) this;
-        }
-    }
-
-    public static class Criteria extends GeneratedCriteria {
-
-        protected Criteria() {
-            super();
-        }
-    }
-
-    public static class Criterion {
-        private String condition;
-
-        private Object value;
-
-        private Object secondValue;
-
-        private boolean noValue;
-
-        private boolean singleValue;
-
-        private boolean betweenValue;
-
-        private boolean listValue;
-
-        private String typeHandler;
-
-        public String getCondition() {
-            return condition;
-        }
-
-        public Object getValue() {
-            return value;
-        }
-
-        public Object getSecondValue() {
-            return secondValue;
-        }
-
-        public boolean isNoValue() {
-            return noValue;
-        }
-
-        public boolean isSingleValue() {
-            return singleValue;
-        }
-
-        public boolean isBetweenValue() {
-            return betweenValue;
-        }
-
-        public boolean isListValue() {
-            return listValue;
-        }
-
-        public String getTypeHandler() {
-            return typeHandler;
-        }
-
-        protected Criterion(String condition) {
-            super();
-            this.condition = condition;
-            this.typeHandler = null;
-            this.noValue = true;
-        }
-
-        protected Criterion(String condition, Object value, String typeHandler) {
-            super();
-            this.condition = condition;
-            this.value = value;
-            this.typeHandler = typeHandler;
-            if (value instanceof List<?>) {
-                this.listValue = true;
-            } else {
-                this.singleValue = true;
-            }
-        }
-
-        protected Criterion(String condition, Object value) {
-            this(condition, value, null);
-        }
-
-        protected Criterion(String condition, Object value, Object secondValue, String typeHandler) {
-            super();
-            this.condition = condition;
-            this.value = value;
-            this.secondValue = secondValue;
-            this.typeHandler = typeHandler;
-            this.betweenValue = true;
-        }
-
-        protected Criterion(String condition, Object value, Object secondValue) {
-            this(condition, value, secondValue, null);
-        }
-    }
-}
